Align Stack and Image Analysis Finalised
