+++ Gat-IJ-Plugin/GAT-Java-Plugin/src/main/java/UI/panes/WorkflowDashboards/AlignStackDashboard.java
@@ -1,109 +1,129 @@
 package UI.panes.WorkflowDashboards;
 
-import ij.IJ;
 import ij.ImagePlus;
 import ij.gui.ImageCanvas;
-import ij.gui.ImageWindow;
-import ij.gui.StackWindow;
-import ij.measure.ResultsTable;
+import ij.gui.Plot;
 
 import javax.swing.*;
+import javax.swing.event.ChangeEvent;
+import javax.swing.event.ChangeListener;
 import java.awt.*;
-import Features.Core.Params;
+import java.io.*;
+import java.util.ArrayList;
+import java.util.List;
 
+/**
+ * Dashboard for displaying aligned image stacks and alignment results.
+ */
 public class AlignStackDashboard extends JPanel {
+
     private final JTabbedPane tabs = new JTabbedPane();
-    private int runCount = 0;
 
     public AlignStackDashboard() {
         super(new BorderLayout());
         add(tabs, BorderLayout.CENTER);
     }
 
-    public void addRun(Params p) {
-        runCount++;
+    /**
+     * Adds a new aligned stack + results CSV to the dashboard.
+     * @param imp The aligned ImagePlus stack to display.
+     * @param resultCSV The results CSV generated by the alignment.
+     */
+    public void addAlignedStackWithResults(ImagePlus imp, File resultCSV) {
+        if (imp == null) return;
 
-        SwingWorker<Void, Void> worker = new SwingWorker<Void, Void>() {
-            private ImagePlus alignedImp;
-            private String resultsText = "";
+        // --- Parse CSV dynamically ---
+        File cleanCSV = new File(resultCSV.getParent(), "CLEANED_" + resultCSV.getName());
+        List<Double> slices = new ArrayList<>();
+        List<Double> dx = new ArrayList<>();
+        List<Double> dy = new ArrayList<>();
 
-            @Override
-            protected Void doInBackground() {
-                // 1) Open image stack
-                alignedImp = IJ.openImage(p.imagePath);
-                if (alignedImp == null) {
-                    JOptionPane.showMessageDialog(AlignStackDashboard.this,
-                            "Could not open image:\n" + p.imagePath,
-                            "Error", JOptionPane.ERROR_MESSAGE);
-                    return null;
-                }
+        try (BufferedReader br = new BufferedReader(new FileReader(resultCSV));
+            PrintWriter pw = new PrintWriter(new FileWriter(cleanCSV))) {
 
-                // 2) Validate reference frame
-                if (p.referenceFrame < 1 || p.referenceFrame > alignedImp.getStackSize()) {
-                    JOptionPane.showMessageDialog(AlignStackDashboard.this,
-                            "Reference frame out of bounds. Image stack has " + alignedImp.getStackSize() + " slices.",
-                            "Error", JOptionPane.ERROR_MESSAGE);
-                    return null;
-                }
+            String headerLine = br.readLine(); // original header
+            if (headerLine == null) return;
 
-                // 3) Create ImageWindow so plugin can render
-                ImageWindow win = new ImageWindow(alignedImp);
-                win.setVisible(false); // hide native ImageJ window
+            String[] headings = headerLine.split(",");
+            int sliceCol = headings.length - 3; // second last column
+            int dxCol = headings.length - 2;    // last column
+            int dyCol = headings.length - 1;
 
-                // 4) Run alignment plugin silently
-                String options = "method=5 windowsizex=32 windowsizey=32 x0=0 y0=0 " +
-                        "swindow=0 subpixel=" + p.subpixel +
-                        " itpmethod=0 ref.slice=" + p.referenceFrame +
-                        (p.useSIFT ? " sift=true" : "") +
-                        " show=true";
-                // make sure correct position selected
-                if (alignedImp != null) {
-                    // call helper from AlignStack instance or duplicate code here
-                    IJ.run(alignedImp, "Align slices in stack...", options);
-                }
+            // write cleaned header
+            pw.println("Slice,Dx,Dy");
 
-                // 5) Capture results
-                ResultsTable rt = ResultsTable.getResultsTable();
-                StringBuilder sb = new StringBuilder();
+            String line;
+            while ((line = br.readLine()) != null) {
+                String[] tokens = line.split(",");
+                if (tokens.length < 3) continue;
 
-                for (int i = 0; i < rt.getCounter(); i++) {
-                    sb.append("Slice: ").append((int)rt.getValue("Slice", i))
-                    .append(" X displacement: ").append(rt.getValue("X Displacement", i))
-                    .append(" Y displacement: ").append(rt.getValue("Y Displacement", i))
-                    .append("\n");
-                }
+                double sliceVal = Double.parseDouble(tokens[sliceCol]);
+                double dxVal = Double.parseDouble(tokens[dxCol]);
+                double dyVal = Double.parseDouble(tokens[dyCol]);
 
-                resultsText = sb.toString();
-                return null;
+                slices.add(sliceVal);
+                dx.add(dxVal);
+                dy.add(dyVal);
+
+                pw.println(sliceVal + "," + dxVal + "," + dyVal);
             }
 
-            @Override
-            protected void done() {
-                // --- Image panel ---
-                ImageCanvas canvas = new ImageCanvas(alignedImp);
-                JScrollPane scroll = new JScrollPane(canvas);
-                scroll.setPreferredSize(new Dimension(900, 600));
-                scroll.setBorder(BorderFactory.createTitledBorder("Aligned Stack Run " + runCount));
+        } catch (IOException ex) {
+            ex.printStackTrace();
+            return;
+        }
 
-                // --- Slice slider ---
-                JSlider sliceSlider = new JSlider(1, alignedImp.getStackSize(), 1);
-                sliceSlider.addChangeListener(e -> {
-                    alignedImp.setSlice(sliceSlider.getValue());
-                    canvas.repaint();
-                });
+        // --- Create Plot ---
+        double[] x = slices.stream().mapToDouble(Double::doubleValue).toArray();
+        double[] yDx = dx.stream().mapToDouble(Double::doubleValue).toArray();
+        double[] yDy = dy.stream().mapToDouble(Double::doubleValue).toArray();
 
-                JPanel panel = new JPanel(new BorderLayout());
-                panel.add(scroll, BorderLayout.CENTER);
-                panel.add(sliceSlider, BorderLayout.SOUTH);
+        Plot plot = new Plot("Alignment Motion", "Slice", "Pixels");
+        plot.setColor(Color.RED);
+        plot.addPoints(x, yDx, Plot.LINE);
+        plot.setColor(Color.BLUE);
+        plot.addPoints(x, yDy, Plot.LINE);
 
-                tabs.addTab("Aligned Run " + runCount, panel);
-                tabs.setSelectedIndex(tabs.getTabCount() - 1);
+        ImagePlus plotImp = plot.getImagePlus();
 
-                revalidate();
-                repaint();
-            }
-        };
+        // --- Aligned stack with slider ---
+        JLabel imageLabel = new JLabel();
+        imageLabel.setHorizontalAlignment(SwingConstants.CENTER);
+        imageLabel.setVerticalAlignment(SwingConstants.CENTER);
+        updateImageLabel(imageLabel, imp, 1);
 
-        worker.execute();
+        JScrollPane imageScroll = new JScrollPane(imageLabel);
+        imageScroll.setPreferredSize(new Dimension(500, 600));
+        imageScroll.setBorder(BorderFactory.createTitledBorder("Aligned Stack"));
+
+        JSlider slider = new JSlider(1, imp.getStackSize(), 1);
+        slider.addChangeListener(e -> updateImageLabel(imageLabel, imp, slider.getValue()));
+
+        JPanel imagePanel = new JPanel(new BorderLayout(5, 5));
+        imagePanel.add(imageScroll, BorderLayout.CENTER);
+        imagePanel.add(slider, BorderLayout.SOUTH);
+
+        // --- Plot panel ---
+        JScrollPane plotScroll = new JScrollPane(new JLabel(new ImageIcon(plotImp.getImage())));
+        plotScroll.setBorder(BorderFactory.createTitledBorder("Motion Plot"));
+        plotScroll.setPreferredSize(new Dimension(500, 600));
+
+        // --- Combine image + plot ---
+        JPanel combined = new JPanel(new GridLayout(1, 2, 10, 10));
+        combined.add(imagePanel);
+        combined.add(plotScroll);
+
+        // --- Add to tabs ---
+        tabs.addTab(imp.getTitle() + " + Results", combined);
+        tabs.setSelectedIndex(tabs.getTabCount() - 1);
     }
+
+    // --- helper method ---
+    private void updateImageLabel(JLabel label, ImagePlus imp, int slice) {
+        imp.setSlice(slice);
+        label.setIcon(new ImageIcon(imp.getProcessor().getBufferedImage()));
+    }
+
+
+
 }
